requirements.txt
# Reddit Fetcher - Required Dependencies

# Reddit API Wrapper
praw>=7.6.0

# Environment Variable Management
python-dotenv>=0.21.0

# Data Validation
pydantic>=1.10.0

# For testing
pytest>=7.0.0
pytest-mock>=3.10.0
